Reorganize the tracking packages.
- The config panels of the overlap tracker goes in
fiji.plugin.trackmate.gui.components.tracker, like the other ones.

- The package fiji.plugin.trackmate.tracking.jaqaman collects all
classes related to the LAP framework in the Jaqaman paper.
Subpackages: costmatrix and costfunctions.

- The manual tracker has its own package:
fiji.plugin.trackmate.tracking.manual
