+++ src/main/java/fiji/plugin/trackmate/gui/components/tracker/JPanelTrackerSettingsMain.java
@@ -60,7 +60,7 @@ import javax.swing.ScrollPaneConstants;
 import javax.swing.SwingConstants;
 
 import fiji.plugin.trackmate.gui.GuiUtils;
-import fiji.plugin.trackmate.tracking.LAPUtils;
+import fiji.plugin.trackmate.tracking.jaqaman.LAPUtils;
 
 public class JPanelTrackerSettingsMain extends javax.swing.JPanel
 {
