+++ src/main/java/fiji/plugin/trackmate/tracking/jaqaman/costfunction/SquareDistCostFunction.java
@@ -19,31 +19,25 @@
  * <http://www.gnu.org/licenses/gpl-3.0.html>.
  * #L%
  */
-package fiji.plugin.trackmate.tracking.costfunction;
+package fiji.plugin.trackmate.tracking.jaqaman.costfunction;
+
+import fiji.plugin.trackmate.Spot;
 
 /**
- * Interface representing a function that can calculate the cost to link a
- * source object to a target object.
- * 
+ * A cost function that returns cost equal to the square distance. Suited to
+ * Brownian motion.
+ *
  * @author Jean-Yves Tinevez - 2014
- * 
- * @param <K>
- *            the type of the sources.
- * @param <J>
- *            the type of the targets.
+ *
  */
-public interface CostFunction< K, J >
+public class SquareDistCostFunction implements CostFunction< Spot, Spot >
 {
 
-	/**
-	 * Returns the cost to link two objects.
-	 * 
-	 * @param source
-	 *            the source object.
-	 * @param target
-	 *            the target object.
-	 * @return the cost as a double.
-	 */
-	public double linkingCost( K source, J target );
+	@Override
+	public double linkingCost( final Spot source, final Spot target )
+	{
+		final double d2 = source.squareDistanceTo( target );
+		return ( d2 == 0 ) ? Double.MIN_NORMAL : d2;
+	}
 
 }
