+++ src/main/java/fiji/plugin/trackmate/tracking/manual/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Classes implementing the manual tracker. This tracker does not do anything.
+ * It just skips the tracking step and as such its implementation is very
+ * minimal.
+ */
+package fiji.plugin.trackmate.tracking.manual;
